requirements.txt
tensorflow
tensorflow_probability
numpy
matplotlib
jupyter
psutil
pyyaml
